BUG-122470 Move Grafana upgrade related changes under v3.1 stack code
Related to AMBARI-25379 Upgrade AMS Grafana version to 6.4.2.
Grafana has been upgraded to v6.4.2 what caused a few configuration changes and version
bump for Ambari Metrics.
Ambari's 'Manage Versions' page shows the bumped up (0.2.0)
version of Ambari Metrics even if a 3.0 stack is chosen. In that case it should show the
old version (0.1.0). To fix this that stack changes should be moved under 3.1 directory
in the stack code.

Change-Id: If1a5baf06bd9f2b5ac8a7ea6504fe4ca3b0261d0
Reviewed-on: https://gerrit.eng.hortonworks.com/r/325081
Reviewed-by: Szilard Antal <santal@cloudera.com>
Tested-by: Zuul <releng@hortonworks.com>